+++ backend/BrewInventory.Application/Contracts/Fermentables/UpdateFermentableRequest.cs
@@ -9,5 +9,5 @@ public record UpdateFermentableRequest(
     string? Type,
     double Color,
     string? GrainCategory,
-    int? Percentage,
+    double? Percentage,
     double Lovibond);
