+++ backend/BrewInventory.Domain/Entities/Fermentable.cs
@@ -9,6 +9,6 @@ public class Fermentable : Inventory
     public FermentableType Type { get; set; }
     public double Color { get; set; }
     public string? GrainCategory { get; set; }
-    public int? Percentage { get; set; }
+    public double? Percentage { get; set; }
     public double Lovibond { get; set; }
 }
